Click can now show movie title
