+++ src/RowRetry.js
@@ -3,15 +3,13 @@ import axios from "./axios";
 
 import Page from './Page.js'
 import './Row.css'
-import YouTube from "react-youtube"
-import movieTrailer from "movie-trailer"
 
 const base_url = "https://image.tmdb.org/t/p/original/";
 
 function RowRetry({ title, fetchURL, isLargeRow }) {
   const [movies, setMovies] = useState([]);
-
   const [Clicker, setClicker] = useState("");
+  const [clickedMovie, setClickedMovie] = useState("");
 
   useEffect(() => {
     async function fetchData() {
@@ -21,24 +19,18 @@ function RowRetry({ title, fetchURL, isLargeRow }) {
     }
     fetchData();
   }, [fetchURL]);
-
-  //you need this for the youtube video size
-  const opts = {
-    height: "390",
-    width: "100%",
-    playerVars: {
-      autoplay: 1,
-    }
-  }
-
-  const handleClick = (movie) => {
+  
+   function handleClick(name) {
     if (Clicker) {
       setClicker('')
+      setClickedMovie(name)
+      console.log(name)
     } else {
-      setClicker("gMAwNk1m4YQ")
+      setClicker("click")
     }
   }
 
+  
 
   return (
     <div className="row">
@@ -49,17 +41,20 @@ function RowRetry({ title, fetchURL, isLargeRow }) {
           <img
             key={movie.id}
             //********/
-            onClick={() => handleClick(movie)} // add the onclick for each movie/show
+            onClick={ () => handleClick(movie?.title || movie?.name || movie?.original_name) } // add the onclick for each movie/show
             //*******/
             className={`row_poster ${isLargeRow && "row_posterLarge"}`} //if its a isLargeRow className is "row_posterLarge"
             src={`${base_url}${isLargeRow ? movie.poster_path : movie.backdrop_path}`} //if it has isLargeRow use movie.backdrop_path else use the other one
             alt={movie.name}
           />
-          
         ))}
+
       </div>
 
-      {Clicker && <Page name={movie.name}/>}
+      <div className="page_contents">
+        {Clicker && <Page name={clickedMovie}/>}
+      </div>
+    
 
     </div>
   );
